Adding a new unit test and the present mpi_comm test is completely broken. We really need to get the serial and parallel testing automated. Not sure how one can work if even tests hang, core-dump or simply do not work.
